// src/types/metadata.ts
export type GenreId =
  | "action"
  | "adventure"
  | "animation"
  | "biography"
  | "comedy"
  | "crime"
  | "documentary"
  | "drama"
  | "family"
  | "fantasy"
  | "history"
  | "horror"
  | "music"
  | "mystery"
  | "romance"
  | "sci-fi"
  | "thriller"
  | "war"
  | "western";

export type CountryId =
  | "AR"
  | "AT"
  | "AU"
  | "BR"
  | "CA"
  | "CN"
  | "DE"
  | "DK"
  | "ES"
  | "FR"
  | "GB"
  | "IE"
  | "IN"
  | "IT"
  | "JP"
  | "KR"
  | "MX"
  | "NO"
  | "NZ"
  | "SE"
  | "US";